REVERTME set algorithm identifier to SHA1
Generated keystores have hard-coded SHA256 algorithm identifier
in the BootKey and BootSignature classes. But Utils.sign()
always signs with SHA1 hashes, and verify_table() in fs_mgr
also expects the verity table to be signed with SHA1.

For now, change BootKey and BootSignature to specify SHA1 to
align with the rest of the code. Revert this once Google fixes
this mess, we would really prefer to use SHA256 across the board,
but don't want to make invasive changes if upstream is also fixing it.

Change-Id: I4104d5fd56708dea1b0292a2fa39e3e45b79aca1
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/229975
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-by: Leibowitz, Michael <michael.leibowitz@intel.com>
Reviewed-by: Ionita, Catalin <catalin.ionita@intel.com>